Modularize routes.ts wave 3: ai-host test + ruleset
server/routes-modules/ai-host-test.ts (96 lines)
  POST /api/admin/ai-host/test — round-trips a one-token prompt
  through the active provider, surfaces the real upstream error.

server/routes-modules/ruleset.ts (96 lines)
  GET/POST /api/admin/ruleset (raw YAML)
  GET/POST /api/admin/ruleset/structured (parsed JSON ↔ YAML)
  Shared helpers: RULESET_FILES whitelist, readYamlFile,
  writeRulesetFile. Every write calls ManagerAgent.flushConfig()
  so live agents pick up the new rules without a process restart.

routes.ts: 1,454 → 1,326 (-128 lines, -9%).
Combined waves 1+2+3: 2,101 → 1,326 (-37%).
